+++ packages/fire_repo/lib/src/fire_repo_mixin/src/fire_repo_store.dart
@@ -1,14 +1,14 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:fire_repo/fire_repo.dart';
 
-abstract class RepoStore<T> {
-  Future<void> store(T entity);
+abstract class RepoSave<T> {
+  Future<void> save(T entity);
 }
 
-mixin FireRepoStore<T, Id extends FireRepoId<C>, C extends FireCollection>
-    implements FireRepo<T, Id, C>, RepoStore<T> {
+mixin FireRepoSave<T, Id extends FireRepoId<C>, C extends FireCollection>
+    implements FireRepo<T, Id, C>, RepoSave<T> {
   @override
-  Future<void> store(T entity) async {
+  Future<void> save(T entity) async {
     await config.entityDoc(entity).set(entity, SetOptions(merge: true));
   }
 }