+++ packages/ssot/lib/ssot.dart
@@ -1,0 +1,3 @@
+library ssot;
+
+export './store.dart';