+++ packages/fire_repo/CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Release
+
+## 0.0.2
+
+* Rename RepoStore to RepoSave
+
+## 0.0.3
+
+* Rename FireRepoStore to FireRepoSave